requirements.txt
fastapi==0.85.1
uvicorn==0.19.0
pymongo==4.3.2
pytest==7.1.3
requests==2.24.0
PyYAML==5.3.1
PyJWT==1.7.1
email-validator==1.1.2